Move TransitivityMatrixGraph
